Clarify how to quote and respond to the target tier policy requirements
Several times, people have seemed unclear on how to respond to some of
the policy requirements, particularly those that just state things the
target developers must *not* do (e.g. not posting to PRs that break the
target). Add a note that such requirements just need acknowledgement,
nothing more.

Make quoting and responding a "must" rather than an "is encouraged to",
since it's easier to review the requirements that way.