[dontmerge]: Change `setenv` on Linux to panic if threaded
See this thread
https://internals.rust-lang.org/t/synchronized-ffi-access-to-posix-environment-variable-functions/

My main goal here is to have this go through all of Rust's CI
and see if any panics turn up.  There are a variety of `set_env()`
calls, but AFAICS they are all early in main.

Then if that works, I'd like to do a crater run so we can gauge
the fallout in the ecosystem of how many places need to be fixed
for this.

(I think we should introduce `try_set_env()`, but this will still
 help us understand the fallout)