Distinguish between `ensure` and `get` dependencies in the DepGraph.
When calling `ensure` on a query, we are checking that the query has been
computed, but we do not use the resulting value.  Since we do not have
access to that value, the caller's behavior must not change.

The `QueryMode` is a single bit, and is stored as most-significant bit
of each dependency in the DepGraph.