Update LLVM with patches for better llvm-cov diagnostics
Cherry-picks
https://github.com/llvm/llvm-project/commit/ee88b8d63e475a75ae525563edfa95f6fcaac83a
and
https://github.com/llvm/llvm-project/commit/126e7611c70ca41782aa851c2bec132607eb8127

These patches to LLVM were added to help debug occasional errors that
cause coverage reporting to fail. Prior to this patch, the only messaging
was that the coverage data was malformed. Hopefully the improved
messaging will help identify the root cause of these errors, when they
arise, so we can make corrections to coverage output from Rust.