Introduce `RUSTC_STATIC_CLANG_RT_PATH` and `RUSTC_STATIC_UNWIND_PATH` envs
The goal of this envirements variable is easy statically link
`libclang_rt` or `libunwind` into rust.

It also introduces a way to easy extend rust build in a way which allows
to inject more static builds for some rare system if it required.

It haven't been documented because it seems as very deep hack and the
user who needs it will discover this hack by reading sources ;)