Return a `String` from `collapsed_doc_value`, not an Option
This has almost no effect in practice, since most places were using `unwrap_or_default`,
and the rest were doing checks that *should* have been checking for an empty string.

The only change in behavior is that the JSON backend no longer distinguishes
"undocumented" and "documented with the empty string". This doesn't seem like a particularly useful distinction,
but I can add it back for that code only if you think it's important.