Approximate Median-of-Medians in select
Previously, there was a cheap approximation to median of medians, but it lead to O(n^2) worst
case behaviour. Instead, use an approximate median of medians algorithm. This is linear time,
but is much cheaper than normal median of medians since it selects random chunks to sort and
only finds the medians of those. It only uses constant space, and in theory could be configured
better to handle small inputs to allocate less.